+++ Marlin/src/HAL/STM32F1/HAL.cpp
@@ -24,58 +24,14 @@
  * HAL for stm32duino.com based on Libmaple and compatible (STM32F1)
  */
 
-#if defined(__STM32F1__) && !defined(VOXELAB_N32)
+#if defined(__STM32F1__)
 
 #include "../../inc/MarlinConfig.h"
 #include "HAL.h"
 
+#if !defined(VOXELAB_N32)
+
 #include <STM32ADC.h>
-
-// ------------------------
-// Types
-// ------------------------
-
-#define __I
-#define __IO volatile
- typedef struct {
-   __I  uint32_t CPUID;                   /*!< Offset: 0x000 (R/ )  CPUID Base Register                                   */
-   __IO uint32_t ICSR;                    /*!< Offset: 0x004 (R/W)  Interrupt Control and State Register                  */
-   __IO uint32_t VTOR;                    /*!< Offset: 0x008 (R/W)  Vector Table Offset Register                          */
-   __IO uint32_t AIRCR;                   /*!< Offset: 0x00C (R/W)  Application Interrupt and Reset Control Register      */
-   __IO uint32_t SCR;                     /*!< Offset: 0x010 (R/W)  System Control Register                               */
-   __IO uint32_t CCR;                     /*!< Offset: 0x014 (R/W)  Configuration Control Register                        */
-   __IO uint8_t  SHP[12];                 /*!< Offset: 0x018 (R/W)  System Handlers Priority Registers (4-7, 8-11, 12-15) */
-   __IO uint32_t SHCSR;                   /*!< Offset: 0x024 (R/W)  System Handler Control and State Register             */
-   __IO uint32_t CFSR;                    /*!< Offset: 0x028 (R/W)  Configurable Fault Status Register                    */
-   __IO uint32_t HFSR;                    /*!< Offset: 0x02C (R/W)  HardFault Status Register                             */
-   __IO uint32_t DFSR;                    /*!< Offset: 0x030 (R/W)  Debug Fault Status Register                           */
-   __IO uint32_t MMFAR;                   /*!< Offset: 0x034 (R/W)  MemManage Fault Address Register                      */
-   __IO uint32_t BFAR;                    /*!< Offset: 0x038 (R/W)  BusFault Address Register                             */
-   __IO uint32_t AFSR;                    /*!< Offset: 0x03C (R/W)  Auxiliary Fault Status Register                       */
-   __I  uint32_t PFR[2];                  /*!< Offset: 0x040 (R/ )  Processor Feature Register                            */
-   __I  uint32_t DFR;                     /*!< Offset: 0x048 (R/ )  Debug Feature Register                                */
-   __I  uint32_t ADR;                     /*!< Offset: 0x04C (R/ )  Auxiliary Feature Register                            */
-   __I  uint32_t MMFR[4];                 /*!< Offset: 0x050 (R/ )  Memory Model Feature Register                         */
-   __I  uint32_t ISAR[5];                 /*!< Offset: 0x060 (R/ )  Instruction Set Attributes Register                   */
-        uint32_t RESERVED0[5];
-   __IO uint32_t CPACR;                   /*!< Offset: 0x088 (R/W)  Coprocessor Access Control Register                   */
- } SCB_Type;
-
-// ------------------------
-// Local defines
-// ------------------------
-
-#define SCS_BASE            (0xE000E000UL)                            /*!< System Control Space Base Address  */
-#define SCB_BASE            (SCS_BASE +  0x0D00UL)                    /*!< System Control Block Base Address  */
-
-#define SCB                 ((SCB_Type       *)     SCB_BASE      )   /*!< SCB configuration struct           */
-
-/* SCB Application Interrupt and Reset Control Register Definitions */
-#define SCB_AIRCR_VECTKEY_Pos              16                                             /*!< SCB AIRCR: VECTKEY Position */
-#define SCB_AIRCR_VECTKEY_Msk              (0xFFFFUL << SCB_AIRCR_VECTKEY_Pos)            /*!< SCB AIRCR: VECTKEY Mask */
-
-#define SCB_AIRCR_PRIGROUP_Pos              8                                             /*!< SCB AIRCR: PRIGROUP Position */
-#define SCB_AIRCR_PRIGROUP_Msk             (7UL << SCB_AIRCR_PRIGROUP_Pos)                /*!< SCB AIRCR: PRIGROUP Mask */
 
 // ------------------------
 // Serial ports
@@ -111,6 +67,50 @@
     }
   #endif
 #endif
+
+// ------------------------
+// ADC
+// ------------------------
+
+static uint16_t adc_results[ADC_COUNT];
+
+// Init the AD in continuous capture mode
+void MarlinHAL::adc_init() {
+  static const uint8_t adc_pins[] = {
+    OPTITEM(HAS_TEMP_ADC_0, TEMP_0_PIN)
+    OPTITEM(HAS_TEMP_ADC_1, TEMP_1_PIN)
+    OPTITEM(HAS_TEMP_ADC_2, TEMP_2_PIN)
+    OPTITEM(HAS_TEMP_ADC_3, TEMP_3_PIN)
+    OPTITEM(HAS_TEMP_ADC_4, TEMP_4_PIN)
+    OPTITEM(HAS_TEMP_ADC_5, TEMP_5_PIN)
+    OPTITEM(HAS_TEMP_ADC_6, TEMP_6_PIN)
+    OPTITEM(HAS_TEMP_ADC_7, TEMP_7_PIN)
+    OPTITEM(HAS_HEATED_BED, TEMP_BED_PIN)
+    OPTITEM(HAS_TEMP_CHAMBER, TEMP_CHAMBER_PIN)
+    OPTITEM(HAS_TEMP_ADC_PROBE, TEMP_PROBE_PIN)
+    OPTITEM(HAS_TEMP_COOLER, TEMP_COOLER_PIN)
+    OPTITEM(HAS_TEMP_BOARD, TEMP_BOARD_PIN)
+    OPTITEM(HAS_TEMP_SOC, TEMP_SOC_PIN)
+    OPTITEM(FILAMENT_WIDTH_SENSOR, FILWIDTH_PIN)
+    OPTITEM(HAS_ADC_BUTTONS, ADC_KEYPAD_PIN)
+    OPTITEM(HAS_JOY_ADC_X, JOY_X_PIN)
+    OPTITEM(HAS_JOY_ADC_Y, JOY_Y_PIN)
+    OPTITEM(HAS_JOY_ADC_Z, JOY_Z_PIN)
+    OPTITEM(POWER_MONITOR_CURRENT, POWER_MONITOR_CURRENT_PIN)
+    OPTITEM(POWER_MONITOR_VOLTAGE, POWER_MONITOR_VOLTAGE_PIN)
+  };
+  static STM32ADC adc(ADC1);
+  // configure the ADC
+  adc.calibrate();
+  adc.setSampleRate((F_CPU > 72000000) ? ADC_SMPR_71_5 : ADC_SMPR_41_5); // 71.5 or 41.5 ADC cycles
+  adc.setPins((uint8_t *)adc_pins, ADC_COUNT);
+  adc.setDMA(adc_results, uint16_t(ADC_COUNT), uint32_t(DMA_MINC_MODE | DMA_CIRC_MODE), nullptr);
+  adc.setScanMode();
+  adc.setContinuous();
+  adc.startConversion();
+}
+
+#endif // !VOXELAB_N32
 
 // ------------------------
 // Watchdog Timer
@@ -172,10 +172,10 @@ void analogWrite(const pin_t pin, int pwm_val8) {
 uint16_t MarlinHAL::adc_result;
 
 // ------------------------
-// Private functions
+// Public functions
 // ------------------------
 
-static void NVIC_SetPriorityGrouping(uint32_t PriorityGroup) {
+void NVIC_SetPriorityGrouping(uint32_t PriorityGroup) {
   uint32_t reg_value;
   uint32_t PriorityGroupTmp = (PriorityGroup & (uint32_t)0x07);               // only values 0..7 are used
 
@@ -186,10 +186,6 @@ static void NVIC_SetPriorityGrouping(uint32_t PriorityGroup) {
                 (PriorityGroupTmp << 8));                                     // Insert write key & priority group
   SCB->AIRCR =  reg_value;
 }
-
-// ------------------------
-// Public functions
-// ------------------------
 
 void flashFirmware(const int16_t) { hal.reboot(); }
 
@@ -221,10 +217,8 @@ void MarlinHAL::init() {
   #if PIN_EXISTS(LED)
     OUT_WRITE(LED_PIN, LOW);
   #endif
-  #if HAS_SD_HOST_DRIVE
+  #ifdef USE_USB_COMPOSITE
     MSC_SD_init();
-  #elif ALL(SERIAL_USB, EMERGENCY_PARSER)
-    usb_cdcacm_set_hooks(USB_CDCACM_HOOK_RX, my_rx_callback);
   #endif
   #if PIN_EXISTS(USB_CONNECT)
     OUT_WRITE(USB_CONNECT_PIN, !USB_CONNECT_INVERTING);  // USB clear connection
@@ -236,124 +230,23 @@ void MarlinHAL::init() {
 
 // HAL idle task
 void MarlinHAL::idletask() {
-  #if HAS_SHARED_MEDIA
-    // If Marlin is using the SD card we need to lock it to prevent access from
-    // a PC via USB.
-    // Other HALs use IS_SD_PRINTING() and IS_SD_FILE_OPEN() to check for access but
-    // this will not reliably detect delete operations. To be safe we will lock
-    // the disk if Marlin has it mounted. Unfortunately there is currently no way
-    // to unmount the disk from the LCD menu.
-    // if (IS_SD_PRINTING() || IS_SD_FILE_OPEN())
-    /* copy from lpc1768 framework, should be fixed later for process HAS_SD_HOST_DRIVE*/
+  #ifdef USE_USB_COMPOSITE
+    #if HAS_SHARED_MEDIA
+      // If Marlin is using the SD card we need to lock it to prevent access from
+      // a PC via USB.
+      // Other HALs use IS_SD_PRINTING() and IS_SD_FILE_OPEN() to check for access but
+      // this will not reliably detect delete operations. To be safe we will lock
+      // the disk if Marlin has it mounted. Unfortunately there is currently no way
+      // to unmount the disk from the LCD menu.
+      // if (IS_SD_PRINTING() || IS_SD_FILE_OPEN())
+      /* copy from lpc1768 framework, should be fixed later for process HAS_SHARED_MEDIA*/
+    #endif
     // process USB mass storage device class loop
     MarlinMSC.loop();
   #endif
 }
 
 void MarlinHAL::reboot() { nvic_sys_reset(); }
-
-// ------------------------
-// Free Memory Accessor
-// ------------------------
-
-extern "C" {
-  extern unsigned int _ebss; // end of bss section
-}
-
-/**
- * TODO: Change this to correct it for libmaple
- */
-
-/*
-// Return free memory between end of heap (or end bss) and whatever is current
-
-#include <wirish/syscalls.c>
-//extern caddr_t _sbrk(int incr);
-#ifndef CONFIG_HEAP_END
-extern char _lm_heap_end;
-#define CONFIG_HEAP_END ((caddr_t)&_lm_heap_end)
-#endif
-
-extern "C" {
-  static int freeMemory() {
-    char top = 't';
-    return &top - reinterpret_cast<char*>(sbrk(0));
-  }
-  int freeMemory() {
-    int free_memory;
-    int heap_end = (int)_sbrk(0);
-    free_memory = ((int)&free_memory) - ((int)heap_end);
-    return free_memory;
-  }
-}
-*/
-
-// ------------------------
-// ADC
-// ------------------------
-
-enum ADCIndex : uint8_t {
-  OPTITEM(HAS_TEMP_ADC_0, TEMP_0)
-  OPTITEM(HAS_TEMP_ADC_1, TEMP_1)
-  OPTITEM(HAS_TEMP_ADC_2, TEMP_2)
-  OPTITEM(HAS_TEMP_ADC_3, TEMP_3)
-  OPTITEM(HAS_TEMP_ADC_4, TEMP_4)
-  OPTITEM(HAS_TEMP_ADC_5, TEMP_5)
-  OPTITEM(HAS_TEMP_ADC_6, TEMP_6)
-  OPTITEM(HAS_TEMP_ADC_7, TEMP_7)
-  OPTITEM(HAS_HEATED_BED, TEMP_BED)
-  OPTITEM(HAS_TEMP_CHAMBER, TEMP_CHAMBER)
-  OPTITEM(HAS_TEMP_ADC_PROBE, TEMP_PROBE)
-  OPTITEM(HAS_TEMP_COOLER, TEMP_COOLER)
-  OPTITEM(HAS_TEMP_BOARD, TEMP_BOARD)
-  OPTITEM(HAS_TEMP_SOC, TEMP_SOC)
-  OPTITEM(FILAMENT_WIDTH_SENSOR, FILWIDTH)
-  OPTITEM(HAS_ADC_BUTTONS, ADC_KEY)
-  OPTITEM(HAS_JOY_ADC_X, JOY_X)
-  OPTITEM(HAS_JOY_ADC_Y, JOY_Y)
-  OPTITEM(HAS_JOY_ADC_Z, JOY_Z)
-  OPTITEM(POWER_MONITOR_CURRENT, POWERMON_CURRENT)
-  OPTITEM(POWER_MONITOR_VOLTAGE, POWERMON_VOLTS)
-  ADC_COUNT
-};
-
-static uint16_t adc_results[ADC_COUNT];
-
-// Init the AD in continuous capture mode
-void MarlinHAL::adc_init() {
-  static const uint8_t adc_pins[] = {
-    OPTITEM(HAS_TEMP_ADC_0, TEMP_0_PIN)
-    OPTITEM(HAS_TEMP_ADC_1, TEMP_1_PIN)
-    OPTITEM(HAS_TEMP_ADC_2, TEMP_2_PIN)
-    OPTITEM(HAS_TEMP_ADC_3, TEMP_3_PIN)
-    OPTITEM(HAS_TEMP_ADC_4, TEMP_4_PIN)
-    OPTITEM(HAS_TEMP_ADC_5, TEMP_5_PIN)
-    OPTITEM(HAS_TEMP_ADC_6, TEMP_6_PIN)
-    OPTITEM(HAS_TEMP_ADC_7, TEMP_7_PIN)
-    OPTITEM(HAS_HEATED_BED, TEMP_BED_PIN)
-    OPTITEM(HAS_TEMP_CHAMBER, TEMP_CHAMBER_PIN)
-    OPTITEM(HAS_TEMP_ADC_PROBE, TEMP_PROBE_PIN)
-    OPTITEM(HAS_TEMP_COOLER, TEMP_COOLER_PIN)
-    OPTITEM(HAS_TEMP_BOARD, TEMP_BOARD_PIN)
-    OPTITEM(HAS_TEMP_SOC, TEMP_SOC_PIN)
-    OPTITEM(FILAMENT_WIDTH_SENSOR, FILWIDTH_PIN)
-    OPTITEM(HAS_ADC_BUTTONS, ADC_KEYPAD_PIN)
-    OPTITEM(HAS_JOY_ADC_X, JOY_X_PIN)
-    OPTITEM(HAS_JOY_ADC_Y, JOY_Y_PIN)
-    OPTITEM(HAS_JOY_ADC_Z, JOY_Z_PIN)
-    OPTITEM(POWER_MONITOR_CURRENT, POWER_MONITOR_CURRENT_PIN)
-    OPTITEM(POWER_MONITOR_VOLTAGE, POWER_MONITOR_VOLTAGE_PIN)
-  };
-  static STM32ADC adc(ADC1);
-  // configure the ADC
-  adc.calibrate();
-  adc.setSampleRate((F_CPU > 72000000) ? ADC_SMPR_71_5 : ADC_SMPR_41_5); // 71.5 or 41.5 ADC cycles
-  adc.setPins((uint8_t *)adc_pins, ADC_COUNT);
-  adc.setDMA(adc_results, uint16_t(ADC_COUNT), uint32_t(DMA_MINC_MODE | DMA_CIRC_MODE), nullptr);
-  adc.setScanMode();
-  adc.setContinuous();
-  adc.startConversion();
-}
 
 void MarlinHAL::adc_start(const pin_t pin) {
   #define __TCASE(N,I) case N: pin_index = I; break;
@@ -386,4 +279,4 @@ void MarlinHAL::adc_start(const pin_t pin) {
   adc_result = (adc_results[(int)pin_index] & 0xFFF) >> (12 - HAL_ADC_RESOLUTION); // shift out unused bits
 }
 
-#endif // __STM32F1__ && !VOXELAB_N32
+#endif // __STM32F1__
